+++ spec/eeny-meeny/experiment_helper_spec.rb
@@ -32,13 +32,13 @@ describe EenyMeeny::ExperimentHelper, experiments: true do
 
           context 'and given a variation id' do
             let(:request_with_variation_cookie) do
-              request.set_cookie(EenyMeeny::Cookie.create_for_experiment_variation(EenyMeeny::Experiment.find_by_id(:my_page), variation_id: :new).to_s)
+              request.set_cookie(EenyMeeny::Cookie.create_for_experiment_variation(EenyMeeny::Experiment.find_by_id(:my_page), :new).to_s)
               request
             end
 
             context 'that matches the variation id the cookie' do
               it "returns the user's experiment variation" do
-                allow(subject).to receive(:cookies).and_return(request_with_cookie.cookie_jar)
+                allow(subject).to receive(:cookies).and_return(request_with_variation_cookie.cookie_jar)
                 expect(subject.participates_in?(:my_page, variation_id: :new)).to be_a EenyMeeny::Variation
                 expect(subject.participates_in?(:my_page, variation_id: 'new')).to be_a EenyMeeny::Variation
               end
@@ -46,7 +46,7 @@ describe EenyMeeny::ExperimentHelper, experiments: true do
 
             context 'that does not match the variation id the cookie' do
               it 'returns nil' do
-                allow(subject).to receive(:cookies).and_return(request_with_cookie.cookie_jar)
+                allow(subject).to receive(:cookies).and_return(request_with_variation_cookie.cookie_jar)
                 expect(subject.participates_in?(:my_page, variation_id: :old)).to be_nil
                 expect(subject.participates_in?(:my_page, variation_id: 'old')).to be_nil
               end